back to f32 for speed and small changes to numericals in diff

+++ examples/diff-debug.rs
@@ -10,13 +10,13 @@ fn main() {
     let face = geom.shapes[0].faces[4].clone();
     println!("face vertices: {:?}", face.data().exterior);
 
-    let m11: Complex<f64> = Complex::new(-0.16685463, -0.013710017);
-    let m12: Complex<f64> = Complex::new(0.958626, 0.078767836);
-    let m21: Complex<f64> = Complex::new(-0.91943324, -0.07554753);
-    let m22: Complex<f64> = Complex::new(-0.16977832, -0.0139502855);
+    let m11: Complex<f32> = Complex::new(-0.16685463, -0.013710017);
+    let m12: Complex<f32> = Complex::new(0.958626, 0.078767836);
+    let m21: Complex<f32> = Complex::new(-0.91943324, -0.07554753);
+    let m22: Complex<f32> = Complex::new(-0.16977832, -0.0139502855);
     let ampl = Matrix2::new(m11, m12, m21, m22);
-    let prop: Vector3<f64> = Vector3::new(-0.6380149, 0.3386723, 0.6915476).normalize();
-    let vk7: Vector3<f64> = Vector3::new(0.38914767, 0.91677344, -0.08994864);
+    let prop: Vector3<f32> = Vector3::new(-0.6380149, 0.3386723, 0.6915476).normalize();
+    let vk7: Vector3<f32> = Vector3::new(0.38914767, 0.91677344, -0.08994864);
     let verts = vec![
         Point3::new(-17.16684, 3.791843, 5.771179),
         Point3::new(-15.896614, 6.7843, 5.7711763),
